+++ include/j_functions.hpp
